First (very rough) draft at new-style PDU decoding. Things compile, but haven't yet been made active (so it's very unlikely they actually work!)   But by checking the changes in now, you can at least see the approach that I'm taking.
git-svn-id: file:///home/hardaker/lib/sf-bkups/net-snmp-convert-svnrepo/branches/NEW-LIBRARY-API@5532 06827809-a52a-0410-b366-d66718629ded

+++ snmplib/net-snmp/struct.h
@@ -121,12 +121,12 @@ struct netsnmp_user_s {
 
 struct netsnmp_pdu_s {
           /* Incomplete, and subject to change.... */
-    int                  version;
-    int                  command;
-    int                  errindex;
-    int                  errstatus;
-    int                  request;
-    int                  flags;
+    long                 version;
+    long                 command;
+    long                 errindex;
+    long                 errstatus;
+    long                 request;
+    long                 flags;
     netsnmp_varbind     *varbind_list;
     netsnmp_comminfo    *community;
     netsnmp_v3info      *v3info;
